% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurosat.R
\name{eurosat_index}
\alias{eurosat_index}
\title{Load EuroSAT MS data index.}
\usage{
eurosat_index()
}
\value{
\link[data.table:data.table]{data.table::data.table} containing the index of EuroSAT files.
}
\description{
This function loads the index of EuroSAT MS data files. To use
this function it is required to have all EuroSAT data downloaded.
}
\note{
Make sure you have executed \link{eurosat_download} before run this
function.
}
\references{
Eurosat: A novel dataset and deep learning benchmark for land use and land
cover classification. Patrick Helber, Benjamin Bischke, Andreas Dengel, Damian
Borth. IEEE Journal of Selected Topics in Applied Earth Observations and
Remote Sensing, 2019.
}
\seealso{
\link{eurosat_download} to download EuroSAT (MS) data.
}
\author{
Felipe Carlos, \email{efelipecarlos@gmail.com}
}
